+++ R/bhm.R
@@ -14,9 +14,11 @@
 #' @param formula an object of class "\link{formula}": a description of which variable
 #' holds the energy readouts and which variable holds the daily temperatures.
 #' @param data a data frame in which the energy and daily temperatures are to be found.
+#' @param baseLoad a optional constant base load, e.g. for domestic hot water preparation.
 #' @return \code{bhm} returns an object of \link{class} "\code{bhm}". The generic
-#' accessor functions \code{coefficients} and \code{residuals} extract the usual
-#' information from the fitted model.
+#' accessor functions \code{coefficients}, \code{vcov} and \code{residuals} extract the usual
+#' information from the fitted model, while \code{logposterior} will return a function
+#' that evaluates the log-posterior as a function of the parameters.
 #' @examples 
 #' set.seed(1111)
 #' 
@@ -43,16 +45,18 @@
 #' coef(fit2)
 #' resid(fit2)
 #' @export
-bhm <- function(formula, data) {
+bhm <- function(formula, data, baseLoad = NULL) {
   Q <- energy(formula, data)
   stopifnot(all(Q >= 0))
   T <- temperatures(formula, data)
-  fit <- posteriorMode(Q, T, minusLogPosterior)
   result <- list()
-  result$coefficients <- stats4::coef(fit)
+  result$logPosterior <- logPosteriorFunction(Q, T)
+  fit <- posteriorMode(result$logPosterior, baseLoad)
+  stopifnot('DHW' %in% names(fit$par) || !is.null(baseLoad))
+  result$coefficients <- c(fit$par, DHW = baseLoad)
   result$residuals <- with(as.list(result$coefficients),
                            Q - epochEnergy(K, tb, DHW, T))
-  result$vcov <- stats4::vcov(fit)
+  result$vcov <- solve(fit$hessian)
   class(result) <- "bhm"
   result
 }
@@ -67,26 +71,28 @@ temperatures <- function(formula, data) {
   data[[attr(terms, "term.labels")]]
 }
 
-posteriorMode <- function(Q, T, logd) {
-  fit <- stats4::mle(logd(Q, T), method = "Nelder")
-  # restart once
-  fit <- stats4::mle(logd(Q, T), method = "Nelder",
-                     start = as.list(fit@coef), control = list(maxit = 1000))
+posteriorMode <- function(logposterior, baseLoad = NULL) {
+  obj <- function(par, ...) {
+   - do.call(logposterior, as.list(c(par, ...)))
+  }
+  start <- list(K = 1, tb = 10, DHW = 1, sigma = 1)
+  fixed <- if (!is.null(baseLoad)) list(DHW = baseLoad) else list()
+  start[names(fixed)] <- NULL
+  fit <- stats::optim(start, obj, gr = NULL, fixed, method = "SANN", control = list(maxit = 20000))
+  fit <- stats::optim(fit$par, obj, gr = NULL, fixed, control = list(maxit = 1000), hessian = TRUE)
   fit
 }
 
-minusLogPosterior <- function(energy, temperatures) {
-  function(K = 1, tb = 1, DHW = 1, sigma = 1) {
-    if (sigma <= 0 || tb < 0 || K <= 0)
-      Inf
-    else {
-      predictedEnergy <- epochEnergy(K, tb, DHW, temperatures)
-      stopifnot(length(predictedEnergy) == length(energy))
-      epochLength <- sapply(temperatures, length)
-      (1 + length(energy)) * log(sigma) + 
-        chisquare(energy, predictedEnergy, sqrt(epochLength) * sigma) / 2
-    }
+logPosteriorFunction <- function(energy, temperatures) {
+  mlp <- function(K = 10, tb = 20, DHW = 0, sigma = 100) {
+    if (sigma <= 0) return(-Inf)
+    predictedEnergy <- epochEnergy(K, tb, DHW, temperatures)
+    stopifnot(length(predictedEnergy) == length(energy))
+    epochLength <- sapply(temperatures, length)
+    - ((1 + length(energy)) * log(sigma) + 
+         chisquare(energy, predictedEnergy, sqrt(epochLength) * sigma) / 2)
   }
+  Vectorize(mlp)
 }
 
 chisquare <- function(energy, predictedEnergy, epochSigma) {
@@ -106,7 +112,17 @@ dailyEnergy <- function(K, tb, DHW, temperature) K * pos(tb - temperature) + DHW
 pos <- function(x) pmax(x, 0)
 
 #' @export
-residuals.bhm <- function(bhm) bhm$residuals
+residuals.bhm <- function(object, ...) object$residuals
 
 #' @export
-vcov.bhm <- function(bhm) bhm$vcov
+vcov.bhm <- function(object, ...) object$vcov
+
+#' Log-posterior of a Bayesian Heating Model
+#' 
+#' Provides the log-posterior of a heating model given the data, as a function
+#' of the model's parameters.
+#' 
+#' @param bhm a fitted model returned by a call to \code{bhm()}
+#' @return a function of the model's parameters (currently K, tb, DHW and sigma)
+#' @export
+logposterior <- function(bhm) bhm$logPosterior